홀리스 끝 밥먹으러 감

+++ 숙제/3개월차/팀 과제(2차-록맨)/8일차-라인 에디터, 저장, 불러오기, 스크롤/한영웅_블록 에디터/BoxEditor/BlockMgr.cpp
@@ -391,7 +391,7 @@ void CBlockMgr::Load_Block()
 
 		bool bResult = ReadFile(hFile,
 								pBuffer,
-								ReadSize,
+								(DWORD)ReadSize,
 								&dwByte, nullptr);
 
 		dwTotalByte += dwByte;
